code/requirements.txt
Jinja2
PyQt5
PyQt5-Qt5
PyQt5_sip
PyQtWebEngine
PyQtWebEngine-Qt5
PyYAML
Pygments
QtAwesome
QtPy
python-dotenv
mysql
mysql-connector-python
mysqlclient
numpy
pyasyncore
pysnmp
python-nmap
pyvis
xmltodict